Avoid error when clicking sort on users_count column
When the sort icon is clicked on the USERS column of the UserGroup list an error is returned. Subsequent requests to this page will result an in error because it is saved as the users preference to sort by this column. Setting the sortable attribute to false for this column resolves this problem.