Bind `rainlab.user.register` event in Plugin.php
I came across this while troubleshooting an issue wherein the `new_user` email was not being sent upon registration, despite having the Notify plugin installed and a corresponding notification rule created.  After making this change locally, the problem was resolved, leading me to believe that this may have been a simple oversight when the change to the Notify plugin was made.